localization system & blacklisted words back !
localization system is finished, but still need to finish all translation.

+++ core/commands/main/ano.js
@@ -1,18 +1,15 @@
 const { MessageEmbed } = require("discord.js");
 
-exports.launch = (client, message, args) => {
+exports.launch = (client, message, args, lang) => {
   const anoMsg = args.join(' ');
+  if (!anoMsg) return;
+
   const embed = new MessageEmbed()
-    .setAuthor("Anonyme", "https://i.imgur.com/xUhKRUT.png")
+    .setAuthor(lang.author, "https://i.imgur.com/xUhKRUT.png")
     .setColor(0x2F3136)
     .setDescription(anoMsg)
     .setTimestamp()
     .setFooter("gokium", client.user.displayAvatarURL({format: "png" || "gif"}));
 
   return message.delete().then(() => message.channel.send(embed));
-}
-
-exports.commands = {
-  description: 'Envoie un message en anonyme.',
-  use: "ano"
 }
